Major upgrade: Mines Wikipedia News for inputs, writes charts,
INPUT
This version gives the option of going to wikipedia’s news today to
choose the topics
OR
Producing a json file with manually selected news topics and pages.
This replaces the hard-coded input pages. Yay!

OUTPUT
This version writes CSV files and image files for each group of news
pages.

CLEANING AND ANALYSIS:
The rollup functions fix a bug which incorrectly missed Wall St Journal
blogs.

The rest of the changes (log files, data and outputs) are artifacts of
the functionality alterations.